Move connect fragment to fragment package
